+++ categories.go
@@ -6,7 +6,6 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"sort"
@@ -51,7 +50,7 @@ func categoriesHandler(packagesBasePaths []string, cacheTime time.Duration) func
 
 		cacheHeaders(w, cacheTime)
 		jsonHeader(w)
-		fmt.Fprint(w, string(data))
+		w.Write(data)
 	}
 }
 
@@ -208,5 +207,5 @@ func getCategoriesOutput(ctx context.Context, categories map[string]*Category) (
 		outputCategories = append(outputCategories, c)
 	}
 
-	return json.MarshalIndent(outputCategories, "", "  ")
+	return util.MarshalJSONPretty(outputCategories)
 }